requirements.txt
torch>=1.7.0
numpy>=1.19.0
python-chess>=1.3.0
